# public/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "SenseiSeek API",
    "version": "1.1.0"
  },
  "security": [
    {
      "cookieAuth": []
    }
  ],
  "paths": {
...
    "/api/shortlist": {
      "post": {
        "summary": "Toggle shortlist for an executive",
        "requestBody": {
          "description": "Shortlist toggle payload",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShortlistToggle"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}